requirements.txt
asgiref
certifi
Django
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
idna
iniconfig
oauthlib
packaging
pluggy
pycparser
PyJWT
pytest
pytest-django
python-decouple
pytz